Create Fibonacci
The Fibonacci series is a sequence of numbers formed by adding each number with the previous one. In this series, which continues in this way, the golden ratio emerges when the numbers are proportional to each other, that is, when a number is divided by the number before it, a series that approaches the golden ratio is obtained.